pyproject.toml
[tool.poetry]
name = "picasso"
version = "0.1.0"
description = "test"
authors = ["Nikita <ldc.endoscopy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.6"
python-dotenv = "^1.0.0"
celery = "^5.3.4"
redis = "^5.0.1"
djangorestframework = "^3.14.0"
flowers = "^0.1.3"
django-celery-results = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

